enable agent hover again, (it is still bad on most viewers for opensim

+++ OpenSim/Region/ClientStack/Linden/Caps/AgentPreferencesModule.cs
@@ -77,11 +77,11 @@ namespace OpenSim.Region.ClientStack.LindenCaps
             {
                 RegisterCaps(agentID, caps);
             };
-/*
+
             ISimulatorFeaturesModule simFeatures = scene.RequestModuleInterface<ISimulatorFeaturesModule>();
             if(simFeatures != null)
                 simFeatures.AddFeature("AvatarHoverHeightEnabled",OSD.FromBoolean(true));
-*/
+
         }
 
         public void PostInitialise() {}
@@ -153,8 +153,7 @@ namespace OpenSim.Region.ClientStack.LindenCaps
             }
             if (req.ContainsKey("hover_height"))
             {
-                //data.HoverHeight = (float)req["hover_height"].AsReal();
-                data.HoverHeight = req["hover_height"].AsReal();
+                data.HoverHeight = (float)req["hover_height"].AsReal();
             }
             if (req.ContainsKey("language"))
             {
@@ -177,10 +176,10 @@ namespace OpenSim.Region.ClientStack.LindenCaps
             resp["hover_height"] = data.HoverHeight;
             resp["language"] = data.Language;
             resp["language_is_public"] = data.LanguageIsPublic;
-/*
+
             IAvatarFactoryModule afm = m_scenes[0].RequestModuleInterface<IAvatarFactoryModule>();
             afm?.SetPreferencesHoverZ(agent, (float)data.HoverHeight);
-*/
+
             string response = OSDParser.SerializeLLSDXmlString(resp);
             return response;
         }
